+++ templates/index.html
@@ -13,7 +13,7 @@
         justify-content: center;
         height: 100vh;
       }
-    
+
       .btnGroup {
         display: flex;
         flex-direction: column;
@@ -21,7 +21,7 @@
         justify-content: center;
         height: 100vh;
       }
-    
+
       .btn {
         background-color: #4caf50;
         border: none;
@@ -34,29 +34,27 @@
         margin: 4px 2px;
         cursor: pointer;
       }
-    
+
       .btn:hover {
         background-color: #3e8e41;
       }
-    
+
       .info {
         background-color: #2196f3;
         color: white;
       }
-    
+
       h1 {
         font-size: 3rem;
         font-weight: 400;
       }
     </style>
-    
   </head>
   <body>
     <h1>Welcome to ConvoAnalytics</h1>
     <div class="btnGroup">
       <a href="/register_form" class="btn">Register</a>
-      <a href="/login_form" class="btn info">Login</a>
+      <a href="/login" class="btn info">Login</a>
     </div>
   </body>
 </html>
-
